+++ client/components/index.js
@@ -3,8 +3,11 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
+
 export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
+export {default as LandingHome} from './landing-home'
 export {default as UserData} from './user-data'
 export {default as ImageUploadForm} from './image-upload-form'
+export {default as NeuronSketch} from './neuron-sketch'
 export {Login, Signup} from './auth-form'