+++ client/index.js
@@ -1,5 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+
+// import 'bootstrap/dist/css/bootstrap.css';
+import 'bootstrap/dist/css/bootstrap.min.css'
+
 import {Provider} from 'react-redux'
 import {Router} from 'react-router-dom'
 import history from './history'